chore: refactor name of js basic
